AI-RL/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyodbc
SQLAlchemy
pandas